+++ rollup.config.ts
@@ -46,8 +46,10 @@ export default defineConfig([
   {
     input: 'src/index.ts',
     output: {
-      file: 'dist/index.d.ts',
+      dir: 'dist',
       format: 'esm',
+      preserveModules: true,
+      preserveModulesRoot: 'src',
     },
     plugins: [dtsPlugin()],
   },